Split FabienCrassatCurriculumVitaeExtensionTest file
